// frontend/src/api/schemas/srcApiConfigurationsTraceWithDevice.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Prism Platform API
 * API for managing traces, configurations, devices, and queries
 * OpenAPI spec version: 1.0.0
 */
import type { SrcApiConfigurationsTraceWithDeviceHostName } from './srcApiConfigurationsTraceWithDeviceHostName';

export interface SrcApiConfigurationsTraceWithDevice {
  trace_id: string;
  trace_name: string;
  trace_timestamp: string;
  trace_filename: string;
  device_id: string;
  device_name: string;
  host_name?: SrcApiConfigurationsTraceWithDeviceHostName;
}